refactor: decompose api main.go, remove dead code, fix stale tests
Split the 975-line cmd/api/main.go (one ~900-line main()) into focused
files in the same package, with no behavior change -- all 17 routes,
K8s Job specs, env var names, notification text, and DB schema preserved.

- cmd/api/config.go: loadConfig() + env-override helpers (envStr/envTrue/
  applyCodebaseEnv), collapsing ~110 lines of override boilerplate
- cmd/api/server.go: Server struct with handlers as methods; extracts
  buildLauncher (was duplicated 2x) and parseWebhook/parsedHook (merges
  the near-identical GitLab/Codeup branches, keeping Codeup's projectId
  fallback chain)
- cmd/api/notify.go: sendNotifications helper (was duplicated 3x)
- main.go: now slim assembly + graceful shutdown only

Fix stale failing tests: source_url_test.go's 5 Codeup cases asserted the
old duplicate-orgId output that commit d6944cf already corrected in
BuildSourceUrl. Updated expectations; suite is green again.

Remove dead code: Repository.GetJobPods, RunnerConfig.PipelineUrl,
Config.Salt (+ its env override), and the unused renderStatusSolid JS
function. Fix log.Fatalf(err.Error()) to log.Fatal(err) in runner.go.

Update Makefile api/api-linux targets to build cmd/api/*.go (the package
now spans multiple files; single-file build would fail).
