+++ Makefile
@@ -8,7 +8,7 @@ pre_build:
 
 # macOS builds
 api:
-	CGO_ENABLED=0 go build -trimpath -ldflags="-s -w" -o $(BUILD_DIR)/neutron-api cmd/api/main.go
+	CGO_ENABLED=0 go build -trimpath -ldflags="-s -w" -o $(BUILD_DIR)/neutron-api cmd/api/*.go
 	chmod a+x $(BUILD_DIR)/neutron-api
 gitlab:
 	CGO_ENABLED=0 go build -trimpath -ldflags="-s -w" -o $(BUILD_DIR)/neutron-gitlab-runner cmd/gitlab-runner/*.go
@@ -19,7 +19,7 @@ codeup:
 
 # Linux cross-compile (for Docker images)
 api-linux:
-	GOOS=linux GOARCH=arm64 CGO_ENABLED=0 go build -trimpath -ldflags="-s -w" -o $(BUILD_DIR)/neutron-api-linux cmd/api/main.go
+	GOOS=linux GOARCH=arm64 CGO_ENABLED=0 go build -trimpath -ldflags="-s -w" -o $(BUILD_DIR)/neutron-api-linux cmd/api/*.go
 gitlab-linux:
 	GOOS=linux GOARCH=arm64 CGO_ENABLED=0 go build -trimpath -ldflags="-s -w" -o $(BUILD_DIR)/neutron-gitlab-runner-linux cmd/gitlab-runner/*.go
 codeup-linux:
